[core] Do not create TorrentManager's directly
EngineSettings contains the path where torrent metadata
files are cached when torrents are downloaded using
MagnetLinks.

If you create a TorrentManager you must register it
with a ClientEngine before it can be used to do
anything.

As such, it's probably easier to combine this into
one simple API call: ClientEngine.AddAsync.

This API call will create the TorrentManager and
register it with the engine in one go. It's one
fewer method calls to accomplish the same task.

Additionally, it now makes it much easier to manage
how/where torrent metadata files are cached and
loaded from the cache.